==> 00_cron.R <==
source('/home/jbontruk/BitBucket/smart_price/01_get_api_v2.R')
source('/home/jbontruk/BitBucket/smart_price/02_gather_data.R')
source('/home/jbontruk/BitBucket/smart_price/03_create_excel_mvp.R')
source('/home/jbontruk/BitBucket/smart_price/04_send_email.R')
